+++ extension/src/classes/broker.hpp
@@ -47,8 +47,8 @@ class Broker : public FixedHold {
     virtual void buy_cargo(const TownCargo* cargo);
 
     int sell_cargo(int type, int amount);
-    int add_cargo(int type, int amount) override;
-    int add_cargo_ignore_accepts(int type, int amount) override;
+    float add_cargo(int type, float amount) override;
+    float add_cargo_ignore_accepts(int type, float amount) override;
     void report_change_in_cash(float amount);
 
     virtual void place_order(int type, int amount, bool buy, float maxPrice);
